float32 for inputs to match model weights

+++ rl.py
@@ -36,7 +36,7 @@ def run_rl(policy : SFNN,
         for t in range(1000):
             # Insert SFNN policy here
             # action = env.action_space.sample() # Random policy
-            action = policy(torch.tensor(observation, dtype=torch.float16), torch.tensor(reward, dtype=torch.float16))
+            action = policy(torch.tensor(observation, dtype=torch.float32), torch.tensor(reward, dtype=torch.float32))
             observation, reward, terminated, truncated, _ = env.step(action.item())
             
             episode_reward += reward  # Accumulate reward
